fix: update meal section
