+++ package.json
@@ -35,7 +35,7 @@
         "@vitejs/plugin-vue": "^2.3.3",
         "@vue/server-renderer": "^3.2.31",
         "@vueuse/core": "^8.5.0",
-        "atmosphere-ui": "^0.7.4",
+        "atmosphere-ui": "^0.7.5",
         "autoprefixer": "^10.4.7",
         "axios": "^0.27.2",
         "date-fns": "^2.22.1",
