--- resources/js/Pages/Planner.vue
@@ -1,185 +1,0 @@
-<template>
-    <AppLayout>
-        <template #header>
-            <div class="flex justify-between">
-                <div>
-                    <h2 class="text-xl font-semibold leading-tight text-pink-400">
-                        Meal Planner
-                    </h2>
-                </div>
-
-                <div class="space-x-2">
-                    <AtButton class="h-10 text-white bg-pink-400" @click="openRandomModal()" rounded> Random Meal </AtButton>
-                    <AtButton class="h-10 text-white bg-pink-400" @click="toggleMode()" rounded>
-                        {{ toggleBtnText }}
-                    </AtButton>
-                </div>
-            </div>
-        </template>
-
-        <div class="py-12">
-            <div class="px-6 mx-auto max-w-7xl md:px-0">
-                <AtDatePager
-                    class="h-12 mb-10 bg-slate-600 border-slate-700"
-                    v-model="date"
-                    v-model:dateSpan="week"
-                    next-mode="week"
-                />
-
-                <div v-if="isGroceryList" class="py-5 overflow-hidden border rounded-md bg-slate-600">
-                    <div v-for="(ingredient, name) in ingredients" class="px-5 text-pink-500 cursor-pointer bg-slate-600">
-                        {{name }} ({{ ingredient.quantity }}) {{ ingredient.unit }}
-                    </div>
-                </div>
-
-                <div v-else class="pt-5 overflow-hidden text-gray-200 border divide-y-2 rounded-md divide-slate-800 border-slate-800 bg-slate-600">
-                    <div v-for="day in week" :key="day" @click="openDayInModal(day)" class="px-5 py-4 cursor-pointer bg-slate-600">
-                        {{ getDayName(day) }}
-
-                        <div class="">
-                            <div v-for="meal in getDayMeals(day)" :key="meal.id" class="font-bold text-pink-400">
-                                {{ meal.dateable.name }}
-                            </div>
-                        </div>
-                    </div>
-                </div>
-
-                <PlanModal
-                    :show="isModalOpen"
-                    :closeable="true"
-                    :date="isModalOpen"
-                    :selected="selectedMealsOfDay"
-                    :title="`Add a new meal to ${isModalOpen && getDayName(isModalOpen)}`"
-                    :meals="meals"
-                    @close="isModalOpen=false"
-                    @saved="onSaved"
-                />
-
-                <RandomMealModal
-                    :show="isRandomModalOpen"
-                    :closeable="true"
-                    @close="isRandomModalOpen=false"
-                />
-            </div>
-        </div>
-    </AppLayout>
-</template>
-
-<script>
-    import AppLayout from '@/Layouts/AppLayout';
-    import { format } from "date-fns";
-    import { AtButton, AtDatePager } from "atmosphere-ui";
-    import MealSection from '@/Components/Meal';
-    import PlanModal from '../Components/PlanModal.vue';
-    import { Inertia } from '@inertiajs/inertia';
-    import { computed, nextTick, reactive, toRefs, onMounted } from "vue";
-    import RandomMealModal from '../Components/RandomMealModal.vue';
-
-    export default {
-        components: {
-            AppLayout,
-            MealSection,
-            AtButton,
-            PlanModal,
-            AtDatePager,
-            RandomMealModal
-        },
-        props: {
-            mealPlans: {
-                type: Object,
-                required: true
-            },
-            meals: {
-                type: Array,
-                required: true
-            },
-            ingredients: {
-                type: Object,
-                default() {
-                    return {}
-                }
-            },
-            mode: {
-                type: String,
-                default: ''
-            }
-        },
-        setup(props) {
-            const state = reactive({
-                isModalOpen: false,
-                isRandomModalOpen: false,
-                date: new Date(),
-                week: null,
-                selectedMealsOfDay: [],
-                isGroceryList: computed(() => {
-                    return props.mode == 'grocery-list';
-                }),
-                toggleBtnText: computed(() => {
-                    return state.isGroceryList ? 'Meal planner' : 'Grocery List';
-                })
-            })
-
-            const openRandomModal = () => {
-                state.isRandomModalOpen = true
-            }
-
-            const getDayName = (date) => {
-                return format(date, 'iiii')
-            }
-
-            const getDayMeals = (date) => {
-                const isoDate = format(date, 'yyyy-MM-dd');
-                return props.mealPlans.data.filter(mealPlan => mealPlan.date == isoDate);
-            }
-
-            const onSaved = () => {
-                nextTick(() => {
-                    Inertia.reload();
-                })
-            }
-
-            const getMode = (toggled) => {
-                if (toggled) {
-                    return state.isGroceryList ? '' : 'mode=grocery-list';
-                } else {
-                    return props.mode ? `mode=${props.mode}` : '';
-                }
-            }
-
-            const getDates = () => {
-                const startDate = state.week[0];
-                const endDate = state.week[6];
-                return `filter[date]=${format(startDate, 'yyyy-MM-dd')}~${format(endDate, 'yyyy-MM-dd')}`;
-            }
-
-            const toggleMode = () => {
-                Inertia.visit(`/meal-planner?${getMode(true)}`);
-            }
-
-            const onWeekChanged = () => {
-                const params = [getMode(), getDates()].join('&');
-                Inertia.visit(`/meal-planner?${params}`);
-            };
-
-            const openDayInModal = (day) => {
-                state.selectedMealsOfDay = getDayMeals(day).map( item => ({
-                    ...item.dateable,
-                    schedule_id: item.id
-                })
-                );
-                state.isModalOpen = day;
-            }
-
-            return {
-                ...toRefs(state),
-                openRandomModal,
-                getDayName,
-                getDayMeals,
-                onSaved,
-                toggleMode,
-                onWeekChanged,
-                openDayInModal,
-            }
-        }
-    }
-</script>
